+++ client-fabric/CHANGELOG.md
@@ -5,10 +5,10 @@
 
 ### Features
 
-* add Fabric group screen controls ([#3](https://github.com/Vanilla-Game/voicechat-group-tools/issues/3)) ([bb383b7](https://github.com/Vanilla-Game/voicechat-group-tools/commit/bb383b7d3d8925e6629f3ff550a05c4dc36dcfd2))
+* add Fabric group screen controls ([#3](https://github.com/Vanilla-Game/simple-voice-chat-group-management/issues/3)) ([bb383b7](https://github.com/Vanilla-Game/simple-voice-chat-group-management/commit/bb383b7d3d8925e6629f3ff550a05c4dc36dcfd2))
 
 
 ### Bug Fixes
 
-* support Simple Voice Chat 2.6.18 and newer ([#8](https://github.com/Vanilla-Game/voicechat-group-tools/issues/8)) ([ae35a02](https://github.com/Vanilla-Game/voicechat-group-tools/commit/ae35a02e986f5ec380fd12aa4f17adf2f5fbfcbc))
-* support Simple Voice Chat 2.6.22 ([#7](https://github.com/Vanilla-Game/voicechat-group-tools/issues/7)) ([0fef734](https://github.com/Vanilla-Game/voicechat-group-tools/commit/0fef7342b5ee09372a47d9e85c3b3dbdec2d4955))
+* support Simple Voice Chat 2.6.18 and newer ([#8](https://github.com/Vanilla-Game/simple-voice-chat-group-management/issues/8)) ([ae35a02](https://github.com/Vanilla-Game/simple-voice-chat-group-management/commit/ae35a02e986f5ec380fd12aa4f17adf2f5fbfcbc))
+* support Simple Voice Chat 2.6.22 ([#7](https://github.com/Vanilla-Game/simple-voice-chat-group-management/issues/7)) ([0fef734](https://github.com/Vanilla-Game/simple-voice-chat-group-management/commit/0fef7342b5ee09372a47d9e85c3b3dbdec2d4955))